Moved the claim circle logic to its own command
